@@ -1,10 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { StrictMode } from "react";
+
+import { createRoot } from "react-dom/client";
+
 import App from "./components/App";
 
-ReactDOM.render(
-  <React.StrictMode>
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(
+  <StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
