@@ -1,5 +1,7 @@
 import { Link } from "react-router-dom";
 
+import { memo } from "react";
+
 import arrow from "../../../assets/images/icons/arrow.svg";
 import edit from "../../../assets/images/icons/edit.svg";
 import trash from "../../../assets/images/icons/trash.svg";
@@ -7,7 +9,7 @@ import trash from "../../../assets/images/icons/trash.svg";
 import { Card, ListHeader } from "./styles";
 import PropTypes from "prop-types";
 
-export default function ContactsList({
+function ContactsList({
   filteredContacts,
   orderBy,
   onDeleteContact,
@@ -73,3 +75,5 @@ ContactsList.defaultProps = {
   onToggleOrderBy: () => {},
   onDeleteContact: () => {},
 };
+
+export default memo(ContactsList);
